+++ CHANGELOG.md
@@ -118,6 +118,7 @@ As always, a backup of the current system files and settings is performed during
  - Specify package versions for pypi dependencies
  - Update python libraries
  - Add unit testing for Custom Functions
+ - Add ability to change theme from Config dropdown menu
 
 
 ## 8.9.2 (2021-03-16)